TechTreeコンポーネントをリファクタリング
templateには技術エリアと技術ツリーのコンポーネントだけを残す形にしました。
